Restrict logins (and thus registrations) to whitelist
For the initial alpha release, the developer console will only accept
apps from a select few developers. This is to allow features to be
added, bugs to be ironed out, and sufficient infrastructure to be
acquired before mass release.

+++ devconsole/auth/github.go
@@ -81,30 +81,43 @@ func GitHubCallback(c *gin.Context) {
 		return
 	}
 
-	// Add session
-	if err := db.DeleteExpiredSessions(); err != nil {
-		_ = c.AbortWithError(http.StatusInternalServerError, err)
-		return
-	}
-	sessionID, err := db.CreateSession(*user.ID, token.AccessToken)
+	// Check against registration whitelist
+	canRegister, err := db.CanUserRegister(*user.ID)
 	if err != nil {
 		_ = c.AbortWithError(http.StatusInternalServerError, err)
 		return
 	}
 
-	c.SetSameSite(http.SameSiteStrictMode)
-	c.SetCookie(SessionCookie, sessionID, 24*60*60, "/", "", true, true) // Max-Age 1 day
+	if canRegister {
+		// Add session
+		if err := db.DeleteExpiredSessions(); err != nil {
+			_ = c.AbortWithError(http.StatusInternalServerError, err)
+			return
+		}
+		sessionID, err := db.CreateSession(*user.ID, token.AccessToken)
+		if err != nil {
+			_ = c.AbortWithError(http.StatusInternalServerError, err)
+			return
+		}
 
-	registered, reviewer, err := db.GetUserRoles(*user.ID)
-	if err != nil {
-		_ = c.AbortWithError(http.StatusInternalServerError, err)
-		return
+		c.SetSameSite(http.SameSiteStrictMode)
+		c.SetCookie(SessionCookie, sessionID, 24*60*60, "/", "", true, true) // Max-Age 1 day
+
+		registered, reviewer, err := db.GetUserRoles(*user.ID)
+		if err != nil {
+			_ = c.AbortWithError(http.StatusInternalServerError, err)
+			return
+		}
+
+		c.JSON(http.StatusOK, gin.H{
+			"logged_in":  true,
+			"registered": registered,
+			"reviewer":   reviewer,
+			"publisher":  *user.ID == conf.SignerGitHubID,
+		})
+	} else {
+		c.JSON(http.StatusOK, gin.H{
+			"logged_in": false,
+		})
 	}
-
-	c.JSON(http.StatusOK, gin.H{
-		"logged_in":  true,
-		"registered": registered,
-		"reviewer":   reviewer,
-		"publisher":  *user.ID == conf.SignerGitHubID,
-	})
 }
